# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Watheq API's Docs",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "eng.asgamal@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/auth/order": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Order a lawyers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/OrderResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/orderDetails": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "List order details by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "Client order id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/OrderResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/order/laywersList": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "List laywers related to order request",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "Client order id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Optional Page Number",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Optional limit of results | Ignored if page is not set",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Category": {
            "properties": {
                "name": {
                    "type": "integer",
                    "example": "Powers of attorney"
                },
                "discription": {
                    "type": "string",
                    "example": [
                        "Powers of attorney description"
                    ]
                },
                "cost": {
                    "type": "string",
                    "example": "400"
                }
            },
            "type": "object",
            "xml": {
                "name": "Category"
            }
        },
...
        "OrderRequest": {
            "properties": {
                "categoryId": {
                    "type": "integer",
                    "example": "3"
                },
                "delivery": {
                    "type": "string",
                    "enum": [
                        "home",
                        "office"
                    ]
                },
                "clientName": {
                    "type": "string",
                    "example": "Mostafa shady"
                },
                "representativeName": {
                    "type": "string",
                    "example": "Khaled senawy"
                },
                "clientNationalID": {
                    "type": "string",
                    "example": "1458787878"
                },
                "representativeNationalID": {
                    "type": "string",
                    "example": "18587887878"
                },
                "letterNumber": {
                    "type": "string",
                    "example": "1533"
                },
                "letterDate": {
                    "type": "string",
                    "example": "12-6-2018"
                },
                "marriageDate": {
                    "type": "string",
                    "example": "12-6-2017"
                },
                "marriageTime": {
                    "type": "string",
                    "example": "15:60"
                },
                "address": {
                    "type": "string",
                    "example": "23 alex st."
                },
                "time": {
                    "type": "string",
                    "example": "5 hours"
                },
                "distance": {
                    "type": "string",
                    "example": "15 kilometer"
                },
                "latitude": {
                    "type": "string",
                    "example": "31.217951"
                },
                "longitude": {
                    "type": "string",
                    "example": "29.942465"
                }
            },
            "type": "object",
            "xml": {
                "name": "OrderRequest"
            }
        },
        "OrderResponses": {
            "properties": {
                "status": {
                    "type": "string",
                    "example": [
                        "New"
                    ]
                },
                "cost": {
                    "type": "string",
                    "example": [
                        "120"
                    ]
                },
                "delivery": {
                    "type": "string",
                    "enum": [
                        "home",
                        "office"
                    ]
                },
                "clientName": {
                    "type": "string",
                    "example": "Mostafa shady"
                },
                "representativeName": {
                    "type": "string",
                    "example": "Khaled senawy"
                },
                "clientNationalID": {
                    "type": "string",
                    "example": "1458787878"
                },
                "representativeNationalID": {
                    "type": "string",
                    "example": "18587887878"
                },
                "letterNumber": {
                    "type": "string",
                    "example": "1533"
                },
                "letterDate": {
                    "type": "string",
                    "example": "12-6-2018"
                },
                "marriageDate": {
                    "type": "string",
                    "example": "12-6-2017"
                },
                "marriageTime": {
                    "type": "string",
                    "example": "15:60"
                },
                "address": {
                    "type": "string",
                    "example": "23 alex st."
                },
                "time": {
                    "type": "string",
                    "example": "5 hours"
                },
                "distance": {
                    "type": "string",
                    "example": "15 kilometer"
                },
                "clientLat": {
                    "type": "string",
                    "example": "40.7324319"
                },
                "clientLong": {
                    "type": "string",
                    "example": "42.7324319"
                },
                "category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Category"
                    },
                    "xml": {
                        "name": "Category",
                        "wrapped": true
                    }
                },
                "lawyer": {
                    "type": "array",
                    "items": {},
                    "xml": {
                        "name": "LawyerShortData",
                        "wrapped": true
                    }
                },
                "isInAcceptLawyerPeriod": {
                    "type": "integer",
                    "enum": [
                        1,
                        0
                    ]
                },
                "created_at": {
                    "type": "integer",
                    "example": 1511649843
                },
                "accepted_at": {
                    "type": "integer",
                    "example": 1511649843
                },
                "closed_at": {
                    "type": "integer",
                    "example": 15511649843
                }
            },
            "type": "object",
            "xml": {
                "name": "OrderResponses"
            }
        },
...
    },
...
}